backend/requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-nvidia-ai-endpoints>=0.1.2
openai>=1.0.0
supabase>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
google-auth>=2.29.0
google-api-python-client>=2.126.0
gspread>=6.1.0
pypdf>=4.1.0
pdfplumber>=0.11.0
apscheduler>=3.10.0
slack-sdk>=3.27.0
jira>=3.8.0
typing_extensions>=4.0.0
